+++ src/main/java/lotto/model/Rank.java
@@ -1,4 +1,4 @@
-package lotto.constant;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.Comparator;
